[all] Refactor Firebase App / FirestoreAdmin / StorageAdmin (#257)
