+++ CODEOWNERS
@@ -13,5 +13,6 @@ admin/ andras@socialincome.org michael@socialincome.org sandino@socialincome.org
 functions/ andras@socialincome.org michael@socialincome.org sandino@socialincome.org
 shared/ andras@socialincome.org michael@socialincome.org sandino@socialincome.org
 recipients_app/ verena@socialincome.org mikolaj@socialincome.org sandino@socialincome.org
+website/ andras@socialincome.org michael@socialincome.org rene@socialincome.org andino@socialincome.org
 
 *.md sandino@socialincome.org andras@socialincome.org michael@socialincome.org
